refactor(visual-editing): move document reporting to hook
