+++ packages/visual-editing/src/ui/Overlays.tsx
@@ -1,11 +1,5 @@
-import {
-  isAltKey,
-  isHotkey,
-  type SanityNode,
-  type VisualEditingControllerMsg,
-} from '@repo/visual-editing-helpers'
-import {DRAFTS_PREFIX} from '@repo/visual-editing-helpers/csm'
-import type {ClientPerspective, ContentSourceMapDocuments} from '@sanity/client'
+import {isAltKey, isHotkey, type VisualEditingControllerMsg} from '@repo/visual-editing-helpers'
+import type {ClientPerspective} from '@sanity/client'
 import type {Status} from '@sanity/comlink'
 import {
   isHTMLAnchorElement,
@@ -47,6 +41,7 @@ import {overlayStateReducer} from './overlayStateReducer'
 import {PreviewSnapshotsProvider} from './preview/PreviewSnapshotsProvider'
 import {SchemaProvider} from './schema/SchemaProvider'
 import {useController} from './useController'
+import {useReportDocuments} from './useReportDocuments'
 
 const Root = styled.div<{
   $zIndex?: string | number
@@ -73,13 +68,6 @@ function raf2(fn: () => void) {
     if (r0 !== undefined) cancelAnimationFrame(r0)
     if (r1 !== undefined) cancelAnimationFrame(r1)
   }
-}
-
-function isEqualSets(a: Set<string>, b: Set<string>) {
-  if (a === b) return true
-  if (a.size !== b.size) return false
-  for (const value of a) if (!b.has(value)) return false
-  return true
 }
 
 const DocumentReporter: FunctionComponent<{
@@ -241,63 +229,7 @@ export const Overlays: FunctionComponent<{
     return () => unsubs.forEach((unsub) => unsub!())
   }, [comlink, history])
 
-  const lastReported = useRef<
-    | {
-        nodeIds: Set<string>
-        perspective: ClientPerspective
-      }
-    | undefined
-  >(undefined)
-
-  const reportDocuments = useCallback(
-    (documents: ContentSourceMapDocuments, perspective: ClientPerspective) => {
-      comlink?.post({
-        type: 'visual-editing/documents',
-        data: {
-          documents,
-          perspective,
-        },
-      })
-    },
-    [comlink],
-  )
-
-  useEffect(() => {
-    // Report only nodes of type `SanityNode`. Untransformed `SanityStegaNode`
-    // nodes without an `id`, are not reported as they will not contain the
-    // necessary document data.
-    const nodes = elements
-      .map((e) => {
-        const {sanity} = e
-        if (!('id' in sanity)) return null
-        return {
-          ...sanity,
-          id: 'isDraft' in sanity ? `${DRAFTS_PREFIX}${sanity.id}` : sanity.id,
-        }
-      })
-      .filter((s) => !!s) as SanityNode[]
-
-    const nodeIds = new Set<string>(nodes.map((e) => e.id))
-    // Report if:
-    // - Documents not yet reported
-    // - Document IDs changed
-    // - Perspective changed
-    if (
-      !lastReported.current ||
-      !isEqualSets(nodeIds, lastReported.current.nodeIds) ||
-      perspective !== lastReported.current.perspective
-    ) {
-      const documentsOnPage: ContentSourceMapDocuments = Array.from(nodeIds).map((_id) => {
-        const node = nodes.find((node) => node.id === _id)!
-        const {type, projectId: _projectId, dataset: _dataset} = node
-        return _projectId && _dataset
-          ? {_id, _type: type!, _projectId, _dataset}
-          : {_id, _type: type!}
-      })
-      lastReported.current = {nodeIds, perspective}
-      reportDocuments(documentsOnPage, perspective)
-    }
-  }, [elements, perspective, reportDocuments])
+  useReportDocuments(comlink, elements, perspective)
 
   const updateDragPreviewCustomProps = useCallback(
     (x: number, y: number) => {
